Mapping from Noll index i to Zernike indices (u,v)
The file contains the u,v = zernikeIndex(i) function that maps from the
convenient Noll indexing to the Zernike indices. So far I was not able
to detect errors in the mapping up to i = 200 or so. I'm confident it'll
work just fine for larger values as it is not using an approximation.