Added a ZernikeWave class
This class allows, to add/remove/adjust modes to your liking. When you
are all done, hit createWavefront(...) and the final result is ready.
For ease, you can quickly plot the result of the modes or the combined
wavefront from the class.  Finally, the class also allows for
decomposition a wavefront in zernike modes, with
decomposeWavefront(...).